--- src/pages/article/hooks/index.ts
@@ -1,1 +1,0 @@
-export {};
