--- src/pages/article/hooks/queries/index.ts
@@ -1,1 +1,0 @@
-export {};
